Fix publishing workfile
- Unified template export across workfile and templates.
- Fix scene saving per instance.
- Cleanup template loader.
