+++ pype/hosts/harmony/__init__.py
@@ -84,6 +84,51 @@ def ensure_scene_settings():
     harmony.send({"function": func, "args": [valid_settings]})
 
 
+def export_template(backdrops, nodes, filepath):
+    func = """function func(args)
+    {
+        // Add an extra node just so a new group can be created.
+        var temp_node = node.add("Top", "temp_note", "NOTE", 0, 0, 0);
+        var template_group = node.createGroup(temp_node, "temp_group");
+        node.deleteNode( template_group + "/temp_note" );
+
+        // This will make Node View to focus on the new group.
+        selection.clearSelection();
+        selection.addNodeToSelection(template_group);
+        Action.perform("onActionEnterGroup()", "Node View");
+
+        // Recreate backdrops in group.
+        for (var i = 0 ; i < args[0].length; i++)
+        {
+            Backdrop.addBackdrop(template_group, args[0][i]);
+        };
+
+        // Copy-paste the selected nodes into the new group.
+        var drag_object = copyPaste.copy(args[1], 1, frame.numberOf, "");
+        copyPaste.pasteNewNodes(drag_object, template_group, "");
+
+        // Select all nodes within group and export as template.
+        Action.perform( "selectAll()", "Node View" );
+        copyPaste.createTemplateFromSelection(args[2], args[3]);
+
+        // Unfocus the group in Node view, delete all nodes and backdrops
+        // created during the process.
+        Action.perform("onActionUpToParent()", "Node View");
+        node.deleteNode(template_group, true, true);
+    }
+    func
+    """
+    harmony.send({
+        "function": func,
+        "args": [
+            backdrops,
+            nodes,
+            os.path.basename(filepath),
+            os.path.dirname(filepath)
+        ]
+    })
+
+
 def install():
     print("Installing Pype config...")
