+++ pype/plugins/harmony/load/load_template_workfile.py
@@ -14,37 +14,12 @@ class ImportTemplateLoader(api.Loader):
     label = "Import Template"
 
     def load(self, context, name=None, namespace=None, data=None):
-        # Make backdrops from metadata.
-        backdrops = context["representation"]["data"].get("backdrops", [])
-
-        func = """function func(args)
-        {
-            Backdrop.addBackdrop("Top", args[0]);
-        }
-        func
-        """
-        for backdrop in backdrops:
-            harmony.send({"function": func, "args": [backdrop]})
-
         # Import template.
         temp_dir = tempfile.mkdtemp()
         zip_file = api.get_representation_path(context["representation"])
         template_path = os.path.join(temp_dir, "temp.tpl")
         with zipfile.ZipFile(zip_file, "r") as zip_ref:
             zip_ref.extractall(template_path)
-
-        func = """function func(args)
-        {
-            var template_path = args[0];
-            var drag_object = copyPaste.copyFromTemplate(
-                template_path, 0, 0, copyPaste.getCurrentCreateOptions()
-            );
-            copyPaste.pasteNewNodes(
-                drag_object, "", copyPaste.getCurrentPasteOptions()
-            );
-        }
-        func
-        """
 
         func = """function func(args)
         {
